fix and simplify video like / add to playlist button flow
