+++ lib/videos/views/components/add_to_playlist_button.dart
@@ -1,50 +1,94 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:invidious/videos/states/add_to_playlist_button.dart';
+import 'package:flutter_gen/gen_l10n/app_localizations.dart';
+import 'package:invidious/videos/states/add_to_playlist.dart';
 
-import 'add_to_playlist.dart';
+import 'add_to_playlist_dialog.dart';
 
-class VideoAddToPlaylistButton extends StatelessWidget {
-  String? videoId;
+enum AddToPlayListButtonType {
+  appBar,
+  modalSheet;
+}
 
-  VideoAddToPlaylistButton({Key? key, this.videoId}) : super(key: key);
+class AddToPlayListButton extends StatelessWidget {
+  final String videoId;
+  final AddToPlayListButtonType type;
+  final Function? afterAdd;
+
+  const AddToPlayListButton(
+      {super.key, required this.videoId, this.type = AddToPlayListButtonType.appBar, this.afterAdd});
+
+  showAddToPlaylistDialog(BuildContext context) {
+    var cubit = context.read<AddToPlaylistCubit>();
+    AddToPlaylistDialog.showAddToPlaylistDialog(context, playlists: cubit.state.playlists, videoId: videoId,
+        onAdd: (selectedPlaylistId) async {
+      await cubit.saveVideoToPlaylist(selectedPlaylistId);
+      if (afterAdd != null) {
+        afterAdd!();
+      }
+    });
+  }
 
   @override
   Widget build(BuildContext context) {
-    var textTheme = Theme.of(context).textTheme;
     var colors = Theme.of(context).colorScheme;
-    return BlocBuilder<AddToPlaylistButtonCubit, AddToPlaylistButtonState>(
-      builder: (context, _) => Visibility(
-        visible: _.isLoggedIn,
-        child: Stack(
-          children: [
-            IconButton(
-              style: ButtonStyle(padding: MaterialStateProperty.all<EdgeInsets>(EdgeInsets.zero)),
-              onPressed: () => AddToPlaylist.showAddToPlaylistDialog(context, _.videoId!),
-              icon: const Icon(
-                Icons.add,
+    var textTheme = Theme.of(context).textTheme;
+    var locals = AppLocalizations.of(context)!;
+
+    return BlocProvider(
+      create: (BuildContext context) => AddToPlaylistCubit(AddToPlaylistController(videoId)),
+      child: BlocBuilder<AddToPlaylistCubit, AddToPlaylistController>(builder: (context, _) {
+        var cubit = context.read<AddToPlaylistCubit>();
+        return switch (type) {
+          (AddToPlayListButtonType.modalSheet) => Padding(
+              padding: const EdgeInsets.only(right: 16.0),
+              child: Column(
+                mainAxisSize: MainAxisSize.min,
+                children: [
+                  IconButton.filledTonal(
+                      onPressed: () => showAddToPlaylistDialog(context), icon: const Icon(Icons.playlist_add)),
+                  Text(locals.addToPlaylist)
+                ],
               ),
             ),
-            _.playListCount > 0
-                ? Positioned(
-                    top: 1,
-                    right: 1,
-                    child: GestureDetector(
-                      onTap: () => AddToPlaylist.showAddToPlaylistDialog(context, _.videoId!),
-                      child: Container(
-                        padding: const EdgeInsets.all(4),
-                        decoration: BoxDecoration(color: colors.secondaryContainer, shape: BoxShape.circle),
-                        child: Text(
-                          _.playListCount.toString(),
-                          style: textTheme.labelSmall,
-                        ),
+          (AddToPlayListButtonType.appBar) => Row(
+              children: [
+                IconButton(
+                  onPressed: cubit.toggleLike,
+                  icon: _.isVideoLiked ? const Icon(Icons.favorite) : const Icon(Icons.favorite_border),
+                ),
+                Stack(
+                  children: [
+                    IconButton(
+                      style: ButtonStyle(padding: MaterialStateProperty.all<EdgeInsets>(EdgeInsets.zero)),
+                      onPressed: () => showAddToPlaylistDialog(context),
+                      icon: const Icon(
+                        Icons.add,
                       ),
                     ),
-                  )
-                : const SizedBox.shrink()
-          ],
-        ),
-      ),
+                    _.playListCount > 0
+                        ? Positioned(
+                            top: 1,
+                            right: 1,
+                            child: GestureDetector(
+                              onTap: () => showAddToPlaylistDialog(context),
+                              child: Container(
+                                padding: const EdgeInsets.all(4),
+                                decoration: BoxDecoration(color: colors.secondaryContainer, shape: BoxShape.circle),
+                                child: Text(
+                                  _.playListCount.toString(),
+                                  style: textTheme.labelSmall,
+                                ),
+                              ),
+                            ),
+                          )
+                        : const SizedBox.shrink()
+                  ],
+                )
+              ],
+            )
+        };
+      }),
     );
   }
 }
