+++ test/widget_test.dart
@@ -7,6 +7,7 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:invidious/home/views/screens/home.dart';
 
 import 'package:invidious/main.dart';
